Completely revised to workaround bcompiler bug.
The bcompiler extension intentionally loads the script its trying to compile.  Because the original compiler used common libraries to do its work, it will not work for others who use the same libraries.  In order to prevent this sort of scenario, the compiler was switched from an OOP design to a procedural one.  And a simple one at that.